move dag sync logic
